+++ src/libs/actions/ReimbursementAccount/navigation.js
@@ -5,6 +5,9 @@ import Onyx from 'react-native-onyx';
 import * as store from './store';
 import CONST from '../../../CONST';
 import ONYXKEYS from '../../../ONYXKEYS';
+import ROUTES from '../../../ROUTES';
+import Navigation from '../../Navigation/Navigation';
+import BankAccount from '../../models/BankAccount';
 
 const WITHDRAWAL_ACCOUNT_STEPS = [
     {
@@ -95,8 +98,25 @@ function goToWithdrawalAccountSetupStep(stepID, achData) {
     Onyx.merge(ONYXKEYS.REIMBURSEMENT_ACCOUNT, {achData: {...newACHData, ...achData, currentStep: stepID}});
 }
 
+/**
+ * Navigate Bank Account Route
+ *
+ * @param {String} reimbursementAccount
+ * @param {String} policyID
+ */
+function navigateToBankAccountRoute(reimbursementAccount, policyID) {
+    const state = lodashGet(reimbursementAccount, 'achData.state');
+    const isShowPage = lodashGet(reimbursementAccount, 'achData.bankAccountID') && state !== BankAccount.STATE.OPEN;
+    if (isShowPage) {
+        Navigation.navigate(ROUTES.getWorkspaceBankAccountRoute(policyID));
+    } else {
+        Navigation.navigate(ROUTES.getBankAccountRoute());
+    }
+}
+
 export {
     goToWithdrawalAccountSetupStep,
     getNextStepToComplete,
     getNextStepID,
+    navigateToBankAccountRoute,
 };
